chore: add more events
