--- packages/helpers/browser/functions/getFingerprint.ts
@@ -1,6 +1,0 @@
-import { getCurrentBrowserFingerPrint } from '@rajesh896/broprint.js'
-
-export const getFingerprint = async () => {
-  const fingerprint = await getCurrentBrowserFingerPrint()
-  return fingerprint
-}
